Use ARGB colors for common video caption preview

+++ feature/postcomposer/src/commonMain/kotlin/com/quata/feature/postcomposer/videoeditor/PostVideoEditorContent.kt
@@ -441,7 +441,7 @@ private fun CommonCaptionPreviewOverlay(
 }
 
 private fun captionColor(argb: Long): Color =
-    Color(argb.toULong())
+    Color(argb.toInt())
 
 @Composable
 private fun PostVideoEditorControls(
